+++ app/models/concerns/commercial/licence_holder.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Commercial
+  module LicenceHolder
+    extend ActiveSupport::Concern
+
+    included do
+      has_many :licences, class_name: 'Commercial::Licence', dependent: :restrict_with_exception
+    end
+  end
+end
